finance-ml/requirements.txt
pandas==2.0.3
requests==2.31.0
sqlalchemy==2.0.19
pymysql==1.1.0
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
flask==2.3.2
mysql-connector-python==8.1.0
openpyxl==3.1.2